pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "keyrgb"
version = "0.18.6"
description = "Minimal RGB keyboard controller for ITE 8291 keyboards"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "GPL-2.0-or-later"}
authors = [
    {name = "KeyRGB Contributors"}
]
keywords = ["keyboard", "rgb", "ite8291", "backlight", "tuxedo", "wootbook"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: X11 Applications",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: GNU General Public License v2 or later (GPLv2+)",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Hardware",
    "Topic :: Utilities",
]

dependencies = [
    "pystray>=0.19.5",
    "Pillow>=11.0.0",
    "pyusb>=1.3.1",
    "evdev>=1.6.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.14.0",
    "black>=24.0.0",
    "mypy>=1.8.0",
]

qt = [
    "PyQt6>=6.10.0",
]

[project.urls]
Homepage = "https://github.com/Rainexn0b/keyRGB"
Issues = "https://github.com/Rainexn0b/keyRGB/issues"
Repository = "https://github.com/Rainexn0b/keyRGB"

[project.scripts]
keyrgb = "src.tray.entrypoint:main"
keyrgb-perkey = "src.gui.perkey:main"
keyrgb-uniform = "src.gui.windows.uniform:main"
keyrgb-reactive-color = "src.gui.windows.reactive_color:main"
keyrgb-calibrate = "src.gui.calibrator:main"
keyrgb-settings = "src.gui.settings:main"
keyrgb-tcc-profiles = "src.gui.tcc.profiles:main"
keyrgb-diagnostics = "src.core.diagnostics:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "buildpython*"]

[tool.setuptools.package-data]
src = ["*.png", "*.json", "core/*.json", "core/resources/*.json"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
pythonpath = ["."]
addopts = "-v --cov=src --cov-report=html --cov-report=term"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "src/*_legacy.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.black]
line-length = 120
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
pretty = true
show_error_codes = true

[tool.ruff.lint]
per-file-ignores = { "tests/**/*.py" = ["E402"] }
